vetproject/requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-environ==0.8.1
django-filter==21.1